refactor(harness): extract extension agent profile
