+++ tests/coding/test_extension_hooks.py
@@ -55,7 +55,7 @@ def _agent_context():
 def test_hook_dispatcher_pipelines_tool_call_decisions() -> None:
     from loushang.agent.types import BeforeToolCallContext
     from loushang.coding.extensions import LoadedExtension, ToolCallDecision
-    from loushang.coding.extensions.hooks import HookDispatcher
+    from loushang.harness.extensions.agent.hooks import ExtensionToolHookDispatcher
 
     seen: list[tuple[str, object]] = []
 
@@ -73,7 +73,7 @@ def test_hook_dispatcher_pipelines_tool_call_decisions() -> None:
         return ToolCallDecision(block=True, reason="blocked by extension")
 
     runner_diagnostics: list[object] = []
-    dispatcher = HookDispatcher(
+    dispatcher = ExtensionToolHookDispatcher(
         [
             LoadedExtension(
                 name="rewrite",
@@ -114,7 +114,7 @@ def test_hook_dispatcher_pipelines_tool_call_decisions() -> None:
 def test_hook_dispatcher_pipelines_tool_result_decisions() -> None:
     from loushang.agent.types import AfterToolCallContext, AgentToolResult
     from loushang.coding.extensions import LoadedExtension, ToolResultDecision
-    from loushang.coding.extensions.hooks import HookDispatcher
+    from loushang.harness.extensions.agent.hooks import ExtensionToolHookDispatcher
 
     seen: list[tuple[object, object]] = []
 
@@ -138,7 +138,7 @@ def test_hook_dispatcher_pipelines_tool_result_decisions() -> None:
             )
         )
 
-    dispatcher = HookDispatcher(
+    dispatcher = ExtensionToolHookDispatcher(
         [
             LoadedExtension(
                 name="one",
@@ -185,7 +185,7 @@ def test_hook_dispatcher_pipelines_tool_result_decisions() -> None:
 def test_hook_dispatcher_pipelines_explicit_null_tool_result_details() -> None:
     from loushang.agent.types import AfterToolCallContext, AgentToolResult
     from loushang.coding.extensions import LoadedExtension, ToolResultDecision
-    from loushang.coding.extensions.hooks import HookDispatcher
+    from loushang.harness.extensions.agent.hooks import ExtensionToolHookDispatcher
 
     seen: list[tuple[object, object]] = []
 
@@ -203,7 +203,7 @@ def test_hook_dispatcher_pipelines_explicit_null_tool_result_details() -> None:
         seen.append((event.result.details, event.hook_details))
         return None
 
-    dispatcher = HookDispatcher(
+    dispatcher = ExtensionToolHookDispatcher(
         [
             LoadedExtension(
                 name="clear",
@@ -250,7 +250,7 @@ def test_hook_dispatcher_pipelines_explicit_null_tool_result_details() -> None:
 def test_hook_dispatcher_records_tool_hook_errors_and_continues() -> None:
     from loushang.agent.types import BeforeToolCallContext
     from loushang.coding.extensions import LoadedExtension, ToolCallDecision
-    from loushang.coding.extensions.hooks import HookDispatcher
+    from loushang.harness.extensions.agent.hooks import ExtensionToolHookDispatcher
 
     runtime_errors: list[tuple[str, str, str]] = []
 
@@ -263,7 +263,7 @@ def test_hook_dispatcher_records_tool_hook_errors_and_continues() -> None:
         return ToolCallDecision(arguments={"ok": True})
 
     diagnostics: list[object] = []
-    dispatcher = HookDispatcher(
+    dispatcher = ExtensionToolHookDispatcher(
         [
             LoadedExtension(
                 name="broken",
